+++ ng_trajectory/optimizers/matryoshka/main.py
@@ -17,7 +17,7 @@ from ng_trajectory.parameter import ParameterList
 
 from ng_trajectory.log import (
     print0,
-    logv, logvv, logvvv,
+    log, logv, logvv, logvvv,
     logfileGet, logfileName, logfileFlush
 )
 
@@ -583,15 +583,13 @@ def optimize() -> Tuple[float, numpy.ndarray, numpy.ndarray, numpy.ndarray]:
     ##
 
     with FILELOCK:
-        logv (
+        log (
             "solution:%s"
             % str(numpy.asarray(points).tolist()),
-            level = VERBOSITY
         )
-        logv (
+        log (
             "final:%f"
             % final,
-            level = VERBOSITY
         )
 
     return (